+++ R/standardize.R
@@ -149,6 +149,17 @@ standardize.numeric <- function(x,
                                 scale = NULL,
                                 verbose = TRUE,
                                 ...) {
+  # set default - need to fix this, else we don't know whether this
+  # comes from "center()" or "standardize()". Furthermore, data.frame
+  # methods cannot return a vector of NULLs for each variable - instead
+  # they return NA. Thus, we have to treat NA like NULL
+  if (is.null(scale) || is.na(scale)) {
+    scale <- TRUE
+  }
+  if (is.null(center) || is.na(center)) {
+    center <- TRUE
+  }
+
   args <- .process_std_center(x, weights, robust, verbose, reference, center, scale)
 
   # Perform standardization