+++ tests/testthat/test-standardize-data.R
@@ -22,11 +22,11 @@ test_that("standardize.numeric", {
 # standardize factor / Date -----------------------------------------------
 test_that("standardize.numeric", {
   f <- factor(c("c", "a", "b"))
-  expect_equal(standardize(f), f)
+  expect_identical(standardize(f), f)
   expect_equal(standardize(f, force = TRUE), c(1, -1, 0), ignore_attr = TRUE)
 
   d <- as.Date(c("1989/08/06", "1989/08/04", "1989/08/05"))
-  expect_equal(standardize(d), d)
+  expect_identical(standardize(d), d)
   expect_equal(standardize(d, force = TRUE), c(1, -1, 0), ignore_attr = TRUE)
 })
 
@@ -63,18 +63,18 @@ test_that("standardize.data.frame, NAs", {
   x <- standardize(iris)
   expect_equal(head(x$Sepal.Length), c(-0.9163, -1.1588, -1.4013, -1.5226, -1.0376, NA), tolerance = 0.01)
   expect_equal(head(x$Sepal.Width), c(1.0237, -0.151, 0.3189, 0.0839, 1.2586, 1.9635), tolerance = 0.01)
-  expect_equal(mean(x$Sepal.Length), NA_real_)
+  expect_identical(mean(x$Sepal.Length), NA_real_)
 
   x <- standardize(iris, two_sd = TRUE)
   expect_equal(head(x$Sepal.Length), c(-0.4603, -0.5811, -0.7019, -0.7623, -0.5207, NA), tolerance = 0.01)
   expect_equal(head(x$Sepal.Width), c(0.5118, -0.0755, 0.1594, 0.042, 0.6293, 0.9817), tolerance = 0.01)
-  expect_equal(mean(x$Sepal.Length), NA_real_)
+  expect_identical(mean(x$Sepal.Length), NA_real_)
 
 
   x <- standardize(poorman::group_by(iris, .data$Species))
   expect_equal(head(x$Sepal.Length), c(0.2547, -0.3057, -0.8661, -1.1463, -0.0255, NA), tolerance = 0.01)
   expect_equal(head(x$Sepal.Width), c(0.2369, -1.0887, -0.5584, -0.8235, 0.502, 1.2974), tolerance = 0.01)
-  expect_equal(mean(x$Sepal.Length), NA_real_)
+  expect_identical(mean(x$Sepal.Length), NA_real_)
 })
 
 
@@ -86,25 +86,25 @@ test_that("standardize.data.frame, apend", {
   iris$Sepal.Length[c(32, 12, 109, 92, 119, 49, 83, 113, 64, 30)] <- NA
 
   x <- standardize(iris, append = TRUE)
-  expect_equal(colnames(x), c(
+  expect_identical(colnames(x), c(
     "Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width",
     "Species", "Sepal.Length_z", "Sepal.Width_z", "Petal.Length_z",
     "Petal.Width_z"
   ))
   expect_equal(head(x$Sepal.Length_z), c(-0.8953, -1.1385, -1.3816, -1.5032, -1.0169, -0.5306), tolerance = 0.01)
   expect_equal(head(x$Sepal.Width_z), c(1.04, -0.1029, 0.3543, 0.1257, 1.2685, 1.9542), tolerance = 0.01)
-  expect_equal(mean(x$Sepal.Length_z), NA_real_)
+  expect_identical(mean(x$Sepal.Length_z), NA_real_)
 
   x <- standardize(iris, two_sd = TRUE, append = TRUE)
   expect_equal(head(x$Sepal.Length_z), c(-0.4477, -0.5692, -0.6908, -0.7516, -0.5084, -0.2653), tolerance = 0.01)
   expect_equal(head(x$Sepal.Width_z), c(0.52, -0.0514, 0.1771, 0.0629, 0.6343, 0.9771), tolerance = 0.01)
-  expect_equal(mean(x$Sepal.Length_z), NA_real_)
+  expect_identical(mean(x$Sepal.Length_z), NA_real_)
 
 
   x <- standardize(poorman::group_by(iris, .data$Species), append = TRUE)
   expect_equal(head(x$Sepal.Length_z), c(0.2746, -0.2868, -0.8483, -1.129, -0.0061, 1.1168), tolerance = 0.01)
   expect_equal(head(x$Sepal.Width_z), c(0.1766, -1.1051, -0.5924, -0.8487, 0.4329, 1.2019), tolerance = 0.01)
-  expect_equal(mean(x$Sepal.Length_z), NA_real_)
+  expect_identical(mean(x$Sepal.Length_z), NA_real_)
 })
 
 
@@ -116,17 +116,20 @@ test_that("standardize.data.frame, weights", {
 
   expect_equal(
     sqrt(cov.wt(cbind(x, x), w)$cov[1, 1]),
-    attr(standardize(x, weights = w), "scale")
+    attr(standardize(x, weights = w), "scale"),
+    tolerance = 1e-4
   )
   expect_equal(
     standardize(x, weights = w),
-    standardize(data.frame(x), weights = w)$x
+    standardize(data.frame(x), weights = w)$x,
+    tolerance = 1e-4
   )
 
   # name and vector give same results
   expect_equal(
     standardize(mtcars, exclude = "cyl", weights = mtcars$cyl),
-    standardize(mtcars, weights = "cyl")
+    standardize(mtcars, weights = "cyl"),
+    tolerance = 1e-4
   )
 
   d <- poorman::group_by(mtcars, am)
@@ -211,8 +214,8 @@ test_that("un/standardize, matrix", {
   z2 <- scale(x)
 
   expect_equal(z1, z2, ignore_attr = TRUE)
-  expect_equal(unstandardize(z1), x)
-  expect_equal(unstandardize(z2), unstandardize(z1))
+  expect_equal(unstandardize(z1), x, ignore_attr = TRUE)
+  expect_identical(unstandardize(z2), unstandardize(z1))
 })
 
 test_that("unstandardize with reference (data frame)", {
@@ -226,11 +229,11 @@ test_that("unstandardize with reference (data frame)", {
 })
 
 test_that("unstandardize does nothing with characters and factors", {
-  expect_equal(
+  expect_identical(
     unstandardise(c("a", "b")),
     c("a", "b")
   )
-  expect_equal(
+  expect_identical(
     unstandardise(factor(c(1, 2))),
     factor(c(1, 2))
   )
@@ -238,8 +241,29 @@ test_that("unstandardize does nothing with characters and factors", {
 
 # select helpers ------------------------------
 test_that("standardize regex", {
-  expect_equal(
+  expect_identical(
     standardize(mtcars, select = "pg", regex = TRUE),
     standardize(mtcars, select = "mpg")
+  )
+})
+
+# standardize when only providing one of center/scale ---------------
+test_that("standardize when only providing one of center/scale", {
+  x <- 1:10
+  expect_identical(
+    as.vector(datawizard::standardize(x, center = FALSE)),
+    x / sd(x)
+  )
+  expect_identical(
+    as.vector(datawizard::standardize(x, center = 2)),
+    (x - 2) / sd(x)
+  )
+  expect_identical(
+    as.vector(datawizard::standardize(x, scale = FALSE)),
+    as.vector(datawizard::center(x))
+  )
+  expect_identical(
+    as.vector(datawizard::standardize(x, scale = 1.5)),
+    (x - mean(x)) / 1.5
   )
 })